+++ src/menu/story.tsx
@@ -1,13 +1,20 @@
 import React from 'react'
-
 import { storiesOf } from '@storybook/react'
 import { action } from '@storybook/addon-actions'
 import { withKnobs } from '@storybook/addon-knobs'
-import { BellIcon, BubbleIcon, CheckShieldIcon, HomeIcon, NewspaperIcon, TicketIcon } from 'icon'
+
+import Section from 'layout/section/baseSection'
+
+import BellIcon from 'icon/bellIcon'
+import BubbleIcon from 'icon/bubbleIcon'
+import CheckShieldIcon from 'icon/checkShieldIcon'
+import HomeIcon from 'icon/homeIcon'
+import NewspaperIcon from 'icon/newspaperIcon'
+import TicketIcon from 'icon/ticketIcon'
+
 import Text from 'text'
 import Menu from 'menu'
 import ItemChoice, { ItemChoiceStatus } from 'itemChoice'
-import Section from 'layout/section/baseSection'
 
 const stories = storiesOf('Widgets|Menu', module)
 stories.addDecorator(withKnobs)