+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - **[FIX]** Fix `Profile` medium size rendering
 - **[FIX]** Fix `TextDisplay1` scale
+- **[UPDATE]** Import `icons` as default instead of named export for a better code splitting
 - [...]
 
 # v24.1.0 (25/03/2020)